requirements.txt
Flask
openai-whisper
srt